+++ Packages/com.thenathannator.hidrogen/CHANGELOG.md
@@ -5,3 +5,7 @@ All notable changes to HIDrogen will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+
+- Devices with the `Linux` or `SDL` interface types will automatically be removed from the input system.